+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   files that belong to the set of results from a certain HRM job.
 * `hrm_omero.omero.add_annotation_keyvalue()` can be used to add so-called [Map
   Annotations][c4] (key-value pairs) to objects in OMERO.
+* `hrm_omero.misc.printlog()` can be used to push a message to the log and stdout.
 * Unit tests using [pytest][c2] and [pytest-cov][c3] (incomplete).
 
 ### Changes